A Sample of Korean General Social Survey Data, 2023

This is a simple sample of the Korean General Social Survey (KGSS) data from 2023.

Usage

kgss_sample

Format

A data frame with the following variables.

year

a numeric vector communicating the year of the survey

respid

a numeric vector communicating a unique identifier for the respondent

age

a numeric vector communicating the age of the respondent

female

a numeric vector communicating whether the respondent self-identifies as a woman or a man

employed

a dummy variable for whether the respondent is employed or 'not employed'

unived

a dummy variable for whether the respondent has a four-year university degree

netuse

a numeric vector for hours of internet usage for the respondent

ideo

a numeric vector communicating the ideology of the respondent on a 1-5 scale. 1 = "very liberal". 5 = "very conservative"

si_gbh

a numeric vector of the extent to which the respondent considers gender-based hatred (toward both men and women) to be a serious issue. 1 = 'very serious'. 5 = 'not serious at all'

satisfin

a numeric vector communicating the respondent's satisfaction with their financial situation. 1 = 'very dissatisfied'. 5 = 'very satisfied'

fp_mord

a numeric vector for whether the respondent thinks maintaining order is the most important priority for South Korea.

fpcat

a character vector for what the respondent believes is the most important priority for South Korea. Possible values: "Maintain Order", "Fight Rising Prices", "Give People More Say", "Protect Freedom of Speech".

cntryaffq

a character vector for the country to which the respondent feels closest. Possible values include "USA", "China", "North Korea", "Russia", and "Japan".

Details

Missingness is substantial for one reason or the other. The data are complete cases only. It's not problematic for this purpose, but I did want to make a note of it.

Data were created based on the English version of the data made available by the Survey Research Center at Sung Kyun Kwan University.